% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWordset.R
\name{fastPOStagger}
\alias{fastPOStagger}
\title{fastPOStagger}
\usage{
fastPOStagger(wordDF)
}
\arguments{
\item{wordDF}{Dataframe including one column labeled "word" for tagging}
}
\value{
Original dataframe including part of speech columns.
}
\description{
fastPOStagger
}
\examples{
\dontrun{newwords <- data.frame(word = c("cat", "green", "slowly"))
fastPOStagger(newwords)}
}
