% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRadlibs.R
\name{makeRadlibs}
\alias{makeRadlibs}
\title{makeRadlibs}
\usage{
makeRadlibs(phrase, wordset = NA)
}
\arguments{
\item{phrase}{String including any number of the words noun, verb, adjective, adverb, plural, or interjection}

\item{wordset}{Data table of your choosing with columns "word" and "pos" at the minimum. Preferably all lowercase.}
}
\value{
New string replacing the keywords with alternatives. Hopefully funny.
}
\description{
makeRadlibs
}
\examples{
makeRadlibs("not sure if i should verb or verb because its an adjective noun")
}
