% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWordset.R
\name{makeWordset}
\alias{makeWordset}
\title{makeWordset}
\usage{
makeWordset(word_list)
}
\arguments{
\item{word_list}{Vector or column of words to be assigned part of speech}
}
\value{
Dataframe including part of speech of each word as best we can tell.
}
\description{
makeWordset
}
\examples{
word_list <- sample(wordset$word, 100)
newWords <- makeWordset(word_list)
}
