% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSTagger.R
\name{POSTagger}
\alias{POSTagger}
\title{POSTagger}
\usage{
POSTagger(wordDF)
}
\arguments{
\item{wordDF}{Dataframe including one column labeled "word" for tagging}
}
\value{
Original dataframe including part of speech columns.
}
\description{
POSTagger
}
\examples{
\dontrun{newwords <- data.frame(word = c("cat", "green", "slowly"))
POSTagger(newwords)}
}
