% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proper_nouns}
\alias{proper_nouns}
\title{A list of English proper nouns with the classifications.}
\format{A data frame with 747 rows and 2 variables:
\describe{
  \item{word}{string of the actual word}
  \item{pos}{part of speech (aka celebrity, place, etc)}
}}
\source{
\url{https://www.kaggle.com/vered1986/propernames-categories/version/1}
}
\usage{
data(proper_nouns)
}
\description{
A dataset derived from https://www.kaggle.com/vered1986/propernames-categories/version/1.
The words are British focused, and I have adjusted some
classifications to be easier for users to work with.
}
\keyword{datasets}
